rename to combined_visualization_out
