+++ callbacks/combined_visualization_out.py
